fix bug in mdqn td loss function and polish code
